dashboard/requirements.txt
pandas
numpy
matplotlib
plotly
streamli
streamlit-aggrid
mplsoccer